Reusable API client and games service
