+++ src/services/api-client.ts
@@ -5,7 +5,7 @@ export interface FetchDataResponse<T> {
     results: T[];
 }
 
-export default axios.create({
+const axiosInstance =  axios.create({
     /* uncomment to invoke rawg api directly from the client 
     baseURL: "https://api.rawg.io/api", */
     /* uncomment to invoke local api service from the client
@@ -19,3 +19,24 @@ export default axios.create({
     }
     */
 })
+
+class APIClient<T> {
+    resourcePath: string;
+
+    constructor(resourcePath: string) 
+    {
+        this.resourcePath = resourcePath;
+    }
+
+    get(params: any) : Promise<T>
+    {
+        return axiosInstance
+            .get<T>(this.resourcePath, {
+                params: params
+              })
+            .then((res) => res.data);
+
+    }
+}
+
+export default APIClient;
